+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
  - *login*, *logout*, *register*, *registerBusiness*, *isLoggedIn* moved to Auth provider.
  -  *changePassword*, *sendPasswordResetLink* *resetPassword* moved to Password provider. Removed 
  *password* word from methods names (*change*, *sendResetLink*, *reset*).
+ - Boards *getTitleSuggestionsFor* renamed to *titleSuggestionsFor*
 ### Removed:
  - *comment* and *deleteComment* methods in Pins provider. Use Comments provider instead.
 ### Fixed: