+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
  - Boards *getTitleSuggestionsFor* renamed to *titleSuggestionsFor*
  - Interests methods renamed: *getMain* to *main*, *getInfo* to *info*, *getPinsFor* to *pins*
  - Pins methods renamed: *getRelatedPins* to *related*, *userFeed* to *feed*
+ - Topics methods renamed: *getInfo* to *info*, *getPinsFor* to *pins*
+ - User methods renamed: *getUserName* to *username*
  
 ### Removed:
  - *comment* and *deleteComment* methods in Pins provider. Use Comments provider instead.
@@ -105,11 +107,11 @@ All notable changes to this project will be documented in this file.
 
 ## [4.10.1] - 2016-09-13
 ### Fixed:
- - Checks in *getUserName* and *isBanned* methods
+ - Checks in *username* and *isBanned* methods
 
 ## [4.10.0] - 2016-09-12
 ### Added
- - User *isBanned* and *getUserName* methods
+ - User *isBanned* and *username* methods
 
 ## [4.9.0] - 2016-09-11
 ### Added: